seven_tour/backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.16
psycopg2-binary>=2.9,<2.10
django-cors-headers>=4.0,<4.4
Pillow>=9.0,<10.2 # For ImageField
django-filter>=23.5,<24.3 # For filtering in DRF
dj-rest-auth>=5.0.0,<6.0.0 # For authentication APIs
django-allauth>=0.58.0,<0.62.0 # For social auth (Google) and email confirmation etc.
djangorestframework-simplejwt>=5.2.0,<5.4.0 # For JWT support with dj-rest-auth
gunicorn>=20.1.0,<22.0.0 # For production WSGI server
# Add other dependencies as needed, e.g.:
# python-dotenv # To load .env files for settings
# sentry-sdk # For error tracking
# celery # For background tasks
# flower # For monitoring Celery tasks
# redis # For Celery broker/backend or caching
